% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{add}
\alias{add}
\title{Sum of two numbers.}
\usage{
add(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
\code{starter.sum} returns the sum of all the values present in its arguments.
}
\examples{
add(1, 1)
add(10, 1)
\dontrun{
add("a")
}
}
