% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ins.R
\name{make_path_norm}
\alias{make_path_norm}
\title{Normalize Path Add-in}
\usage{
make_path_norm()
}
\value{
normalized path string
}
\description{
Wrapper to execute \code{\link[fs:path_math]{fs::path_norm()}} as a shortcut on
highlighted text. The updated text will be converted in place to a path
normalized for the environment currently in use.
For instance, "\" or "\\\" will be converted to "/" on Windows machines.
See below for process of setting shortcut.
}
\details{
Add keyboard shortcut for \code{\link[=make_path_norm]{make_path_norm()}} in RStudio
\itemize{
\item Install starter, and restart RStudio
\item Select "Tools" --> "Modify Keyboard Shortcuts...".
\item In Search box, type "Make Path Normal".
\item Click in the "Shortcut" column on the "Make Path Normal" row.
\item Press intended shortcut keys (suggested: Ctrl + Shift + /) to set shortcut.
\item NOTE: It is possible to override a previously specified key combination with this selection.
}
}
\examples{
# must be run interactively and RStudio must be installed
if (interactive()) {
  # set a keyboard shortcut for path normalization
  rstudio.prefs::use_rstudio_keyboard_shortcut(
    "Ctrl+Shift+/" = "starter::make_path_norm"
  )
}
}
\seealso{
\link[fs:path_math]{fs::path_norm}
}
