% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{multiply}
\alias{multiply}
\title{Multiply of two numbers}
\usage{
multiply(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The product of \code{x} and \code{y}.
}
\description{
\code{starter.sum} returns the product of all the values present in its arguments.
}
\examples{
multiply(1, 1)
multiply(10, 1)
\dontrun{
multiply("a")
}
}
