% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_rstudio_project.R
\name{open_rstudio_project}
\alias{open_rstudio_project}
\title{Open RStudio Project}
\usage{
open_rstudio_project(path, open = TRUE)
}
\arguments{
\item{path}{folder path}

\item{open}{Logical indicating whether to open new project in fresh RStudio
session}
}
\description{
If path has an \verb{*.Rproj} file, the project is opened in a new RStudio IDE session.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
open_rstudio_project(path = tempfile())
\dontshow{\}) # examplesIf}
}
