% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{project_templates}
\alias{project_templates}
\title{Project templates}
\format{
A named list containing the project templates.
}
\usage{
project_templates
}
\description{
The \code{project_templates} object defines the contents of the project
templates used in \code{create_project()} and \code{use_file()}.
}
\examples{
if (FALSE) {
create_project(
  path = file.path(tempdir(), "Sjoberg New Project"),
  template = project_templates[["analysis"]]
)
}
}
\keyword{datasets}
