% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starter-package.R
\docType{package}
\name{starter-package}
\alias{starter}
\alias{starter-package}
\title{starter: A Starter Kit for New Projects}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

}
\keyword{internal}
