% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starter.R
\docType{package}
\name{starter}
\alias{starter}
\alias{starter-package}
\title{starter: A package to learn how to create a R package.}
\description{
The starter package provides two simple functions:
sum and multiply.
}
\section{Starter functions}{

The starter functions ...
}

