% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starter-package.R
\docType{package}
\name{starter-package}
\alias{starter}
\alias{starter-package}
\title{starter: Starter Kit for New Projects}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Get started with new projects by dropping a skeleton of a new project into a new or existing directory, initialise git repositories, and create reproducible environments with the 'renv' package. The package allows for dynamically named files, folders, file content, as well as the functionality to drop individual template files into existing projects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ddsjoberg/starter}
  \item \url{https://www.danieldsjoberg.com/starter/index.html}
  \item Report bugs at \url{https://github.com/ddsjoberg/starter/issues}
}

}
\author{
\strong{Maintainer}: Daniel D. Sjoberg \email{danield.sjoberg@gmail.com} (\href{https://orcid.org/0000-0003-0862-2018}{ORCID})

Other contributors:
\itemize{
  \item Emily Vertosick [contributor]
}

}
\keyword{internal}
